Refactor settings components into feature-based architecture
Reorganizes 12 settings-related files from scattered type-based directories (components/, hooks/) into cohesive feature-based structure under src/features/settings/. Creates two sub-features: data-settings/ (5 files) and column-config/ (7 files).

This is Migration Story 03 following the project's file organization principles to improve discoverability and maintainability by colocating related functionality.
