+++ src/features/data-tracking/components/tier-stats-config-panel.tsx
@@ -2,10 +2,10 @@ import { ChevronDown, ChevronUp, RotateCcw, Settings } from 'lucide-react'
 import type { UseTierStatsConfigReturn } from '../hooks/use-tier-stats-config'
 import type { TierStatsAggregation } from '../types/tier-stats-config.types'
 import { getFieldDisplayName, canFieldHaveHourlyRate } from '../utils/tier-stats-config'
-import { useColumnSearch } from '../hooks/use-column-search'
-import { useColumnReorder } from '../hooks/use-column-reorder'
-import { SelectedColumnItem } from './selected-column-item'
-import { SearchableColumnPicker } from './searchable-column-picker'
+import { useColumnSearch } from '../../settings/column-config/use-column-search'
+import { useColumnReorder } from '../../settings/column-config/use-column-reorder'
+import { SelectedColumnItem } from '../../settings/column-config/selected-column-item'
+import { SearchableColumnPicker } from '../../settings/column-config/searchable-column-picker'
 import { Button } from '../../../components/ui/button'
 import { InfoBox } from '../../../components/ui/info-box'
 import { EmptyState } from '../../../components/ui/empty-state'
