+++ src/features/settings/column-config/use-field-filter.ts
@@ -1,6 +1,6 @@
 import { useState, useEffect, useMemo } from 'react';
-import type { FieldTrendData } from '../types/game-run.types';
-import { filterFieldTrends, isValidSearchTerm } from '../utils/field-filter';
+import type { FieldTrendData } from '../../data-tracking/types/game-run.types';
+import { filterFieldTrends, isValidSearchTerm } from '../../data-tracking/utils/field-filter';
 
 interface UseFieldFilterOptions {
   debounceMs?: number;
