+++ src/features/settings/data-settings/data-settings.tsx
@@ -1,6 +1,6 @@
 import { Card, CardContent, CardDescription, CardHeader, CardTitle, Button, Alert, AlertDescription } from '../../../components/ui';
 import { CheckCircle, XCircle, Trash2, Database, AlertTriangle, X } from 'lucide-react';
-import { useDataSettings } from '../hooks/use-data-settings';
+import { useDataSettings } from './use-data-settings';
 
 export function DataSettings() {
   const {
